Refactors CustomSchedule.py for scale out of schedules (#2574)
## Motivation

We are about to land a significant number of custom schedules for
various tile sizes and transpose orientations. By using the convention
"a schedule is described in a function" we can reasonably scale the code
out, instead of a chain of if/elif blocks that spans hundreds or
thousands of lines. Initially we keep the existing schedules in
functions inside CustomSchedule.py, but we can refactor as needed.

## Technical Details


## Test Plan

TENSILELITE_CLIENT_ARGS="--build-type Debug --gpu-targets gfx950
--clean" tox -e py39 -- Tensile/Tests -k custom_mainloop_scheduling.yaml
## Test Result

<!-- Briefly summarize test outcomes. -->